+++ packages/xchain-thorchain-query/src/thorchain-query.ts
@@ -723,7 +723,9 @@ export class ThorchainQuery {
           rune: new CryptoAmount(assetToBase(assetAmount(0)), AssetRuneNative),
         }
         return dustValues
-      case 'BTC' || `BCH` || `LTC`:
+      case 'BTC':
+      case `BCH`:
+      case `LTC`:
         // 10k sats
         dustValues = {
           asset: new CryptoAmount(assetToBase(assetAmount(0.0001)), asset),
@@ -755,6 +757,13 @@ export class ThorchainQuery {
         // 1 million sats
         dustValues = {
           asset: new CryptoAmount(assetToBase(assetAmount(0.01)), asset),
+          rune: new CryptoAmount(assetToBase(assetAmount(0)), AssetRuneNative),
+        }
+        return dustValues
+      case 'AVAX':
+        // 0 AVAX
+        dustValues = {
+          asset: new CryptoAmount(assetToBase(assetAmount(0)), asset),
           rune: new CryptoAmount(assetToBase(assetAmount(0)), AssetRuneNative),
         }
         return dustValues